Revert "fs/opendingux: Added libraries from booboo's rootfs for binary compatibility."
This reverts commit 99d57062b7bf99de61e4e7e071c546156917c854.

We don't gain much binary compatibility from shipping two versions:
imagine a legacy application linked to both libpng 1.2 and SDL_image,
while SDL_image on the rootfs is linked to libpng 1.4. When the
application is run, the dynamic linker will load both versions of
libpng and resolve symbols from one of them, which is the wrong version
for either the application or SDL_image.

We decided to attempt binary compatibility by shipping only the old
version rather than both the old and the new version.