Revert "fs/opendingux: inittab: the framebuffer console is now disabled right before calling the frontend menu."
This reverts commit 5e2b6c59cf8f4ace7c1768f7efdee8d4d792b486.

We might need the logging if the frontend fails to start.

I hope we can find a solution that only displays console messages after
the user presses a key and use that for the kernel messages, the init
scripts and the front-end launch, to get a clean boot and still have
enough info to debug if necessary.