makedevs: Let main() return "ret" variable instead of hardcoded 0.
This fixes a build error if the host GCC warns about the unused variable,
which is fatal because of -Werror. Only some GCC versions issue this
warning, which is a drawback of using -Werror.

Also it makes sure that problems during execution of "makedevs" are
actually reflected in the exit code, of course.