opendingux: Add defconfig.
This is based on the config from the buildroot-2010.11 branch.
Compared to that config, the libconfig and liboil packages were disabled.
Also the bwm-ng package was enabled.
Also we are back to GCC 4.5 for now, because I did not want to merge
all the GCC updates. Expect a return to 4.7 on a later buildroot base.